Factory out DXGI factory creation
