+++ wgpu-hal/src/auxil/dxgi/mod.rs
@@ -1,1 +1,4 @@
 pub mod conv;
+pub mod exception;
+pub mod factory;
+pub mod result;
